requirements.txt
aiogram>=3.0.0
pyyaml>=6.0
docker>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
